Refactor timer handling and configuration parsing
- Simplified the calculation of remaining time and the determination of the last timer in `handle_past_schedule_in_progress` and `create_future_timers`.
- Improved logging for timer scheduling and handling.
- Changed the parsing of `hourStart` and `horarioLen` in `parse_horario` to use `valueint` instead of `valuedouble` for better accuracy.
- Updated overlap checking in `sanity_check_profile` to log integer values instead of floating-point for clarity.
- Added a new configuration file `shrimp-config-2026-02-03.json` with detailed sensor and limits settings.
No esp_timers yet.

+++ main/inc/typedef.h
@@ -170,8 +170,8 @@ PV1V
 // ============================================================================
 
 typedef struct {
-    float hourStart;
-    float horarioLen;
+    uint8_t hourStart;
+    uint8_t horarioLen;
     uint8_t pwmDuty;
 } horario_t;
